refactor: move graph component into map
