+++ src/components/Main/Graph/graph.scss
@@ -1,4 +1,4 @@
-@import '../../shared/variables';
+@import '../../../shared/variables';
 
 .graph__chartdiv {
   height: 100%;
@@ -12,12 +12,11 @@
   left: 0;
   top: 0;
   width: 100%;
-  // height: 100vh;
   height: calc(100vh - 100px);
   z-index: 100;
 }
 
-@media(max-width: 768px) {
+@media (max-width: 768px) {
   .full-screen.graph__chartdiv {
     height: 100vh;
   }
